backend/requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
python-dotenv==1.0.1
Flask-Session==0.6.0 